move DexContracts to explorer.services package;
