+++ modules/explorer-core/src/main/scala/org/ergoplatform/explorer/db/repositories/OutputRepo.scala
@@ -7,13 +7,13 @@ import cats.implicits._
 import doobie.free.implicits._
 import doobie.refined.implicits._
 import fs2.Stream
-import org.ergoplatform.explorer.db.DexContracts
 import org.ergoplatform.explorer.db.algebra.LiftConnectionIO
 import org.ergoplatform.explorer.db.syntax.liftConnectionIO._
 import org.ergoplatform.explorer.db.models.Output
 import org.ergoplatform.explorer.db.models.aggregates.{DexSellOrderOutput, ExtendedOutput}
 import org.ergoplatform.explorer.{Address, BoxId, HexString, TokenId, TxId}
 import org.ergoplatform.explorer.db.doobieInstances._
+import org.ergoplatform.explorer.services.DexContracts
 
 /** [[Output]] and [[ExtendedOutput]] data access operations.
   */
